add category

+++ app/Http/Controllers/BlogController.php
@@ -3,20 +3,30 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BlogController extends Controller
 {
     public function index(){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         $blogs = \App\Blog::all();
         return view('admin.blog.index', array('blogs'=>$blogs));    
     }
 
     public function create(){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         $categories = \App\Category::pluck('name', 'id');
         return view('admin.blog.create', array('categories'=>$categories));
     }
     
     public function store(Request $request){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         $blog = \App\Blog::create([
             'title'=> $request->input('title'),
             'content'=> $request->input('content'),
@@ -33,6 +43,9 @@ class BlogController extends Controller
     }
     
     public function update(Request $request, $blogid){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         if ($request->isMethod('post')) {
             \App\Blog::find($blogid)->update([
                 'title'=> $request->input('title'),
@@ -52,6 +65,9 @@ class BlogController extends Controller
     }
 
     public function delete(Request $request, $id){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         \App\Blog::find($id)->delete();
         flash('Blog eliminado!')->success();
         return redirect('admin/blog/');
