+++ app/Http/Controllers/CategoryController.php
@@ -3,19 +3,29 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index(){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         $categories = \App\Category::all();
         return view('admin.category.index', array('categories'=>$categories));    
     }
 
     public function create(){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         return view('admin.category.create');
     }
     
     public function store(Request $request){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         \App\Category::create([
             'name'=> $request->input('name')
         ]);
@@ -24,6 +34,9 @@ class CategoryController extends Controller
     }
     
     public function update(Request $request, $categoryid){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         if ($request->isMethod('post')) {
             \App\Category::find($categoryid)->update([
                 'name'=> $request->input('name')
@@ -39,6 +52,9 @@ class CategoryController extends Controller
     }
 
     public function delete(Request $request, $id){
+        if (!Auth::check()) {
+            return redirect('login/');
+        }
         try {
             \App\Category::find($id)->delete();
             flash('Categoria eliminado!')->success();
